@@ -28,8 +28,7 @@
     "@babel/preset-env": "^7.14.5",
     "babel-jest": "^27.0.2",
     "babel-loader": "^8.2.2",
-    "css-loader": "^5.2.6",
-    "eslint": "^7.29.0",
+    "eslint": "^7.32.0",
     "eslint-config-airbnb-base": "^14.2.1",
     "eslint-config-prettier": "^8.3.0",
     "eslint-plugin-import": "^2.23.4",
@@ -40,7 +39,7 @@
     "husky": "^6.0.0",
     "jest": "^27.0.4",
     "lint-staged": "^11.0.0",
-    "mini-css-extract-plugin": "^1.6.0",
+    "mini-css-extract-plugin": "^1.6.2",
     "prettier": "^2.3.1",
     "webpack": "^5.40.0",
     "webpack-cli": "^4.7.2",
@@ -59,6 +58,12 @@
   },
   "keywords": [],
   "dependencies": {
-    "axios": "^0.21.1"
+    "@babel/plugin-transform-runtime": "^7.14.5",
+    "axios": "^0.21.1",
+    "css-loader": "^6.2.0",
+    "html-loader": "^2.1.2",
+    "sass-loader": "^12.1.0",
+    "style-loader": "^3.2.1",
+    "url-loader": "^4.1.1"
   }
 }
